requirements.txt
# FastAPI
fastapi>=0.110.0
uvicorn[standard]>=0.30.0

# Database
sqlalchemy>=2.0.0
asyncpg>=0.29.0
alembic>=1.13.0
pgvector>=0.2.5
psycopg2>=2.9.11
# Environment
python-dotenv>=1.0.1

# Authentication & security
pyjwt[crypto]>=2.8.0
passlib[bcrypt]>=1.7.4

